Extract query handler into separate TypeScript file
